Add TrollControl, block Harmony 2.2, fix AssetPath
AssetPath used to require that the mod live in a folder with
the same name as the main DLL. It can now live in any folder.

Harmony 2.2 is incompatible with Harmony (redesigned).
I am working on a built-in zoning adjuster, and switching
detours to Harmony, hence Harmony (redesigned) needs to be
supported. Since Harmony 2.2 breaks compatibility, I am
adding here a roadblock to delay implementation of
Harmony 2.2, which is ill conceived, as described in
TransitModBase.cs

It would be unethical for me to the point of maliciousness
to follow Felix Schmidt (boformer, Colossal Order)'s "upgrade"
plan, so I must choose the roadblock which is of all bad choices
the least bad option for both the community who needs/wants this Mod
long term, and for myself so I don't make a colossally unethical
error. See TransitModBase.cs.

The mod will now run locally without being sensitive to the name of
the folder. Fixes #1