Fix build when Unity is not installed.
In order to generate debug symbols (Debug configuration), Unity
Mono is needed to generate MDB files.
In release mode symbols are not generated, but the project assumed
unity is installed. If this assumption was not true, this caused
the build to fail with an error similar to
"[MSBuild]::NormalizeDirectory('') - path cannot be zero length"

This commit corrects detection of Unity.
In Debug configuration, if Unity is not installed, the build will stop
with an error and instructions to install Unity.
In Release configuration, the build continues without caring that Unity
is not installed.

closes #10