PSA: Backup your mods locally, Harmony 2.2 is here
In e206a9d I removed the Harmony 2.2 roadblock, which means 2.2 was released
(irresponsibly), and this means the entire workshop is going through another
destabilization like in Feb-Mar 2021, as mod authors start experimenting with
new Harmony 2.2 features.

This commit changes the description of the mod to the Public Service
Announcement that your mods are about to start breaking as the API for 2.2 is
released in the following days and modders start "updating".

To avoid disruptions, I advise backing up your mods so you can try using
them locally if needed, to bypass the coming stormy weather on the workshop.